requirements.txt
elasticsearch==8.4.1
python_dateutil==2.8.2
requests~=2.28.1